+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 - Initial experimental version
 - Alphabetic and numeric layouts
 - Keyboard controller and scope
+
+## 0.1.0-dev
+
+- Phone and email layouts
+- Additional keys
+- Keyboard customization